% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.distance.R
\name{base.distance}
\alias{base.distance}
\title{Calulate the distance from a US Air Force Base to a service}
\usage{
base.distance(base = NULL, service = "libraries", ...)
}
\arguments{
\item{base}{The name of a USAF base COUNS}

\item{service}{The type of service for which a distance is desired}

\item{...}{Extra stuff}
}
\description{
Calulate the distance from a US Air Force Base to a service
}

