% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libdistData.R
\docType{data}
\name{Base_Locations}
\alias{Base_Locations}
\title{Locations of CONUS USAF Bases}
\format{A \code{data.table} with 5 variables:
\tabular{rlll}{
 [, 1] \tab Base (Long Name)  \tab Character String of the Library name/address/city/state \tab \bold{Categoric}\cr
 [, 2] \tab Base (Short Name) \tab Ordered list distances (in miles) from the base \tab \bold{Numeric}\cr
 [, 3] \tab Base (Abbreviation) \tab Latitude coordinate of the library \tab \bold{Numeric}\cr
 [, 4] \tab Longitude   \tab Longitude coordinate of the library \tab \bold{Numeric}\cr
 [, 5] \tab Latitude \tab Latitude coordinate of the library \tab \bold{Numeric}
 }}
\source{
http://www.publiclibraries.com

Google Maps Geocode API
}
\usage{
Base_Locations
}
\description{
Information about public libraries in each state
}
\keyword{datasets}

