% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseData.R
\name{baseData}
\alias{baseData}
\title{Scrape and format data for US Public Libraries}
\usage{
baseData(state = NULL, geocode = TRUE, ...)
}
\arguments{
\item{state}{Which state}

\item{geocode}{Should geocode data be provided?}

\item{...}{more options for ggmap::geocode}
}
\description{
Scrape and format data for US Public Libraries
}
\details{
This functions scrapes data from http://www.publiclibraries.com,
recording the name, address, zipcode, latitude, and longitude for each
public library in \code{state}. The resulting data set for each state is
stored as a text file in the extdata directory and as a .rda file.  The
resulting data sets may be called using, for example, \code{publicLibs::AK_Libs}
(which returns data for public libraries in Alaska).
}

