% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressData.R
\name{compressData}
\alias{compressData}
\title{Compress library data files}
\usage{
compressData()
}
\description{
This function uses plyr::join to combine the public library data sets for each state into a single dataset for the entire U.S.
}

