% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libdistData.R
\docType{data}
\name{alts_Libs}
\alias{alts_Libs}
\alias{andr_Libs}
\alias{andw_Libs}
\alias{arnl_Libs}
\alias{bckl_Libs}
\alias{beal_Libs}
\alias{blln_Libs}
\alias{brkd_Libs}
\alias{brks_Libs}
\alias{chrl_Libs}
\alias{clmb_Libs}
\alias{cnnn_Libs}
\alias{crch_Libs}
\alias{dovr_Libs}
\alias{dvsm_Libs}
\alias{dyss_Libs}
\alias{edwr_Libs}
\alias{egln_Libs}
\alias{ells_Libs}
\alias{elmn_Libs}
\alias{elsn_Libs}
\alias{fewr_Libs}
\alias{frch_Libs}
\alias{gdfl_Libs}
\alias{gntr_Libs}
\alias{grnf_Libs}
\alias{hckm_Libs}
\alias{hill_Libs}
\alias{hllm_Libs}
\alias{hnsc_Libs}
\alias{hrlf_Libs}
\alias{krtl_Libs}
\alias{kslr_Libs}
\alias{lckl_Libs}
\alias{lghl_Libs}
\alias{lngl_Libs}
\alias{lsan_Libs}
\alias{lttr_Libs}
\alias{luke_Libs}
\alias{mcch_Libs}
\alias{mccn_Libs}
\alias{mcdl_Libs}
\alias{mcgr_Libs}
\alias{mint_Libs}
\alias{mlms_Libs}
\alias{mnth_Libs}
\alias{mody_Libs}
\alias{mxwl_Libs}
\alias{nlls_Libs}
\alias{offt_Libs}
\alias{otis_Libs}
\alias{pope_Libs}
\alias{ptrc_Libs}
\alias{ptrs_Libs}
\alias{rbns_Libs}
\alias{rndl_Libs}
\alias{schr_Libs}
\alias{sctt_Libs}
\alias{shaw_Libs}
\alias{shpp_Libs}
\alias{symj_Libs}
\alias{tnkr_Libs}
\alias{trvs_Libs}
\alias{tynd_Libs}
\alias{usaf_Libs}
\alias{vanc_Libs}
\alias{vndn_Libs}
\alias{whtm_Libs}
\alias{wrgp_Libs}
\title{Distances From Public Libraries to USAF Bases}
\format{A \code{data.table} with 4 variables:
\tabular{rlll}{
 [, 1] \tab location  \tab Character String of the Library name/address/city/state \tab \bold{Categoric}\cr
 [, 2] \tab miles     \tab Ordered list distances (in miles) from the base \tab \bold{Numeric}\cr
 [, 3] \tab latitide \tab Latitude coordinate of the library \tab \bold{Numeric}\cr
 [, 4] \tab longitude   \tab Longitude coordinate of the library \tab \bold{Numeric}
 }}
\source{
http://www.publiclibraries.com

Google Maps Geocode API
}
\usage{
alts_Libs

andr_Libs

andw_Libs

arnl_Libs

bckl_Libs

beal_Libs

blln_Libs

brkd_Libs

brks_Libs

chrl_Libs

clmb_Libs

cnnn_Libs

crch_Libs

dovr_Libs

dvsm_Libs

dyss_Libs

edwr_Libs

egln_Libs

ells_Libs

elmn_Libs

elsn_Libs

fewr_Libs

frch_Libs

gdfl_Libs

gntr_Libs

grnf_Libs

hckm_Libs

hill_Libs

hllm_Libs

hnsc_Libs

hrlf_Libs

krtl_Libs

kslr_Libs

lckl_Libs

lghl_Libs

lngl_Libs

lsan_Libs

lttr_Libs

luke_Libs

mcch_Libs

mccn_Libs

mcdl_Libs

mcgr_Libs

mint_Libs

mlms_Libs

mnth_Libs

mody_Libs

mxwl_Libs

nlls_Libs

offt_Libs

otis_Libs

pope_Libs

ptrc_Libs

ptrs_Libs

rbns_Libs

rndl_Libs

schr_Libs

sctt_Libs

shaw_Libs

shpp_Libs

symj_Libs

tnkr_Libs

trvs_Libs

tynd_Libs

usaf_Libs

vanc_Libs

vndn_Libs

whtm_Libs

wrgp_Libs
}
\description{
Information about public libraries in each state
}
\keyword{datasets}

