% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{mergeLibs}
\alias{mergeLibs}
\title{Megre state datasets together}
\usage{
mergeLibs()
}
\description{
This function uses plyr::join to combine the public library data sets for each state into a single dataset for the entire U.S.
}

