% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AL_Libs}
\alias{AK_Libs}
\alias{AL_Libs}
\alias{AR_Libs}
\alias{AZ_Libs}
\alias{CA_Libs}
\alias{CO_Libs}
\alias{CT_Libs}
\alias{DE_Libs}
\alias{FL_Libs}
\alias{GA_Libs}
\alias{HI_Libs}
\alias{IA_Libs}
\alias{ID_Libs}
\alias{IL_Libs}
\alias{IN_Libs}
\alias{KS_Libs}
\alias{KY_Libs}
\alias{LA_Libs}
\alias{MA_Libs}
\alias{MD_Libs}
\alias{ME_Libs}
\alias{MI_Libs}
\alias{MN_Libs}
\alias{MO_Libs}
\alias{MS_Libs}
\alias{MT_Libs}
\alias{NC_Libs}
\alias{ND_Libs}
\alias{NE_Libs}
\alias{NH_Libs}
\alias{NJ_Libs}
\alias{NM_Libs}
\alias{NV_Libs}
\alias{NY_Libs}
\alias{OH_Libs}
\alias{OK_Libs}
\alias{OR_Libs}
\alias{PA_Libs}
\alias{RI_Libs}
\alias{SC_Libs}
\alias{SD_Libs}
\alias{TN_Libs}
\alias{TX_Libs}
\alias{US_Libs}
\alias{UT_Libs}
\alias{VA_Libs}
\alias{VT_Libs}
\alias{WA_Libs}
\alias{WI_Libs}
\alias{WV_Libs}
\alias{WY_Libs}
\title{U.S. Public Library Data}
\format{A \code{data.frame} with 7 variables:
\tabular{rlll}{
 [, 1] \tab Library Name  \tab Temperature applied to the bond \tab \bold{Categoric}\cr
 [, 2] \tab Address       \tab Relative humidity applied to the bond \tab \bold{Categoric}\cr
 [, 3] \tab City \tab Time the bond was under load \tab \bold{Categoric}\cr
 [, 4] \tab State   \tab Load at which the bond failed \tab \bold{Numeric}\cr
 [, 5] \tab ZIP Code       \tab Relative humidity applied to the bond \tab \bold{Categoric}\cr
 [, 6] \tab Latitide     \tab Time the bond was under load \tab \bold{Categoric}\cr
 [, 7] \tab Longitude   \tab Load at which the bond failed \tab \bold{Numeric}
 }}
\source{
http://www.publiclibraries.com

Google Maps Geocode API
}
\usage{
AL_Libs

AK_Libs

AZ_Libs

AR_Libs

CA_Libs

CO_Libs

CT_Libs

DE_Libs

FL_Libs

GA_Libs

HI_Libs

ID_Libs

IL_Libs

IN_Libs

IA_Libs

KS_Libs

KY_Libs

LA_Libs

ME_Libs

MD_Libs

MA_Libs

MI_Libs

MN_Libs

MS_Libs

MO_Libs

MT_Libs

NE_Libs

NV_Libs

NH_Libs

NJ_Libs

NM_Libs

NY_Libs

NC_Libs

ND_Libs

OH_Libs

OK_Libs

OR_Libs

PA_Libs

RI_Libs

SC_Libs

SD_Libs

TN_Libs

TX_Libs

UT_Libs

VT_Libs

VA_Libs

WA_Libs

WV_Libs

WI_Libs

WY_Libs

US_Libs
}
\description{
Information about public libraries in each state
}
\keyword{datasets}

